#22 first variant Boilerplate reduced by: 1. Inheritance 2. Remove empty intermediate Observe, Changes (optional step) 3. Improve readability by inner classes vs anonymous object (optional step) Pros: without delegates and additional allocations Cons: Inheritance may limit us in future.
