+++ redom/android/src/main/java/com/pushtorefresh/redom/android/view/LinearLayoutImpl.kt
@@ -2,33 +2,20 @@ package com.pushtorefresh.redom.android.view
 
 import com.jakewharton.rxbinding2.view.RxView
 import com.jakewharton.rxrelay2.PublishRelay
-import com.pushtorefresh.redom.api.Component
-import com.pushtorefresh.redom.api.ComponentGroup
-import com.pushtorefresh.redom.api.LinearLayout
+import com.pushtorefresh.redom.api.*
 import com.pushtorefresh.redom.api.LinearLayout.Orientation.Horizontal
 import com.pushtorefresh.redom.api.LinearLayout.Orientation.Vertical
-import com.pushtorefresh.redom.api.View
-import com.pushtorefresh.redom.api.ViewGroup
-import com.pushtorefresh.redom.api.ViewParent
-import com.pushtorefresh.redom.api.toViewStructure
 import io.reactivex.Observable
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
 import io.reactivex.rxkotlin.plusAssign
 
-class LinearLayoutImpl<O : Any>(private val viewParent: ViewParent<O>) : LinearLayout<O> {
+class LinearLayoutImpl<O : Any>(private val viewParent: ViewParent<O>) : LinearLayout<O>, ViewImpl<O, View.Observe, LinearLayout.Change>() {
 
-    private val outputObservables = mutableListOf<Observable<O>>()
     private val views = mutableListOf<View<O, *, *>>()
-    private var observeClicks: PublishRelay<Any>? = null
     private var _orientation: Observable<LinearLayout.Orientation>? = null
 
-    override val observe = object : LinearLayout.Observe {
-        override val clicks: Observable<Any>
-            get() = observeClicks ?: PublishRelay.create<Any>().also {
-                observeClicks = it
-            }
-    }
+    override val observe: View.Observe = ViewObserveImpl()
 
     override val change = object : LinearLayout.Change {
         override var orientation: Observable<LinearLayout.Orientation>
@@ -36,12 +23,6 @@ class LinearLayoutImpl<O : Any>(private val viewParent: ViewParent<O>) : LinearL
             set(value) {
                 _orientation = value
             }
-    }
-
-    override val output = object : View.Output<O> {
-        override fun plusAssign(observable: Observable<O>) {
-            outputObservables += observable
-        }
     }
 
     override fun <Ob : View.Observe, Ch : View.Change, V : View<O, Ob, Ch>> createView(clazz: Class<out V>): V {
@@ -65,7 +46,7 @@ private class LinearLayoutComponent<O : Any>(private val observeClicks: PublishR
                                              override val output: Observable<O>) :
         ComponentGroup<O, android.widget.LinearLayout> {
 
-    override val clazz: Class<out ViewGroup<out Any, out ViewGroup.Observe, out ViewGroup.Change>> = LinearLayout::class.java
+    override val clazz: Class<out ViewGroup<out Any, out View.Observe, out View.Change>> = LinearLayout::class.java
     override val viewStructure = toViewStructure(this)
 
     override fun bind(view: android.widget.LinearLayout): Disposable {
