+++ redom/api/src/main/kotlin/com/pushtorefresh/redom/api/ViewGroup.kt
@@ -1,14 +1,6 @@
 package com.pushtorefresh.redom.api
 
-interface ViewGroup<O : Any, Ob: ViewGroup.Observe, Ch: ViewGroup.Change> : View<O, Ob, Ch>, ViewParent<O> {
-
-    interface Observe : View.Observe {
-
-    }
-
-    interface Change : View.Change {
-
-    }
+interface ViewGroup<O : Any, Ob: View.Observe, Ch: View.Change> : View<O, Ob, Ch>, ViewParent<O> {
 
     override fun build(): ComponentGroup<O, out Any>
 }
